Разделение js на модули
